test cleanup pass 1
